remove openssl system dependency
This change removes the openssl dependency by disabling the default features of
reqwest and enabling the `rustls-tls` feature so we will use
the native rust TLS library without any system dependencies.

This should fix issues where the binaries could not get executed or the
compilation failed if the system of the user doesn't has a proper
OpenSSL library installed.